+++ examples/linear_layout.rs
@@ -1,0 +1,38 @@
+use embedded_graphics_simulator::{
+    BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, Window,
+};
+
+use embedded_graphics::{
+    fonts::{Font6x8, Text},
+    pixelcolor::BinaryColor,
+    prelude::*,
+    style::TextStyleBuilder,
+};
+use embedded_layout::layout::linear::LinearLayout;
+use embedded_layout::prelude::*;
+
+fn main() -> Result<(), core::convert::Infallible> {
+    let mut display: SimulatorDisplay<BinaryColor> = SimulatorDisplay::new(Size::new(64, 48));
+    let output_settings = OutputSettingsBuilder::new()
+        .theme(BinaryColorTheme::OledBlue)
+        .build();
+
+    let display_area = display.display_area();
+
+    let text_style = TextStyleBuilder::new(Font6x8)
+        .text_color(BinaryColor::On)
+        .build();
+
+    LinearLayout::vertical()
+        .with_alignment(horizontal::Center)
+        .add_view(Text::new("Vertical", Point::zero()).into_styled(text_style))
+        .add_view(Text::new("Linear", Point::zero()).into_styled(text_style))
+        .add_view(Text::new("Layout", Point::zero()).into_styled(text_style))
+        .arrange()
+        .align_to(&display_area, horizontal::Center, vertical::Center)
+        .draw(&mut display)
+        .unwrap();
+
+    Window::new("LinearLayout exmaple", &output_settings).show_static(&display);
+    Ok(())
+}